Move rating controllers into "reviewers" namespace
Specs passing
