--- app/controllers/rating/comments_controller.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-class Rating::CommentsController < CommentsController
-  PATH_PARENTS = [:rating]
-end
