--- app/controllers/rating/overview_controller.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-class Rating::OverviewController < Rating::BaseController
-  def index
-  end
-end
