+++ app/controllers/reviewers/comments_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module Reviewers
+ class CommentsController < CommentsController
+    PATH_PARENTS = [:reviewers]
+  end
+end
