+++ app/controllers/reviewers/overview_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+module Reviewers
+ class OverviewController < Reviewers::BaseController
+    def index
+    end
+  end
+end
