--- app/views/rating/overview/index.html.slim
@@ -1,6 +1,0 @@
-= render 'rating/breadcrumbs'
-
-h1 Application Rating
-ul[class="list-unstyled"]
-  li = link_to 'Your Todos', [:rating, :todos]
-  li = link_to 'All Applications', [:rating, :applications]
