+++ app/views/reviewers/applications/edit.html.slim
@@ -1,4 +1,4 @@
-= render 'rating/breadcrumbs'
+= render 'reviewers/breadcrumbs'
 
 h1 = "Application: #{@application.name}"
