+++ app/views/reviewers/shared/_rating_sidebar.html.slim
