+++ app/views/reviewers/todos/index.html.slim
@@ -1,4 +1,4 @@
-= render 'rating/breadcrumbs'
+= render 'reviewers/breadcrumbs'
 
 h1 Todo
 
@@ -14,7 +14,7 @@ table.table
   tbody
     - @todos.each do |todo|
       tr.todo class=(todo.rating&.updated_at ? 'todo--rated' : '')
-        td= link_to(todo.application.team_name, [:rating, todo.application])
+        td= link_to(todo.application.team_name, [:reviewers, todo.application])
         td= todo.eligible? ? icon('check') : icon('ban')
         td.text-right= todo.sign_offs?.first ? icon('check') : icon('ban')
         td.text-left= todo.sign_offs?.second ? icon('check') : icon('ban')
